Adopt one macOS input policy (#43)
## Summary

- remove the four selectable touch compatibility modes and their
persisted setting
- keep ordinary mouse, trackpad, and Magic Mouse input unchanged
- always repair native macOS double-clicks with the two touch taps
ArenaNet's Wasm client recognizes
- preserve unrelated legacy profile values while retiring touchMode on
the next save

## Verification

- pnpm check
- pnpm exec playwright test tests/electron/input.spec.ts
- pnpm verify

The full gate passed with 575 unit, 123 policy, 20 integration, 21
release, and 91 Electron tests; one opt-in live test was skipped.
Packaging and both packaged runtime smokes passed.

## Stack

This draft targets codex/data-protection-keychain intentionally. After
#42 merges, rebase this branch onto main, retarget the PR, and mark it
ready.

The separate click-to-move investigation is intentionally out of scope.
