+++ src/renderer/index.html
@@ -344,14 +344,6 @@
         <section class="settings-pane" id="settings-pane-advanced" role="tabpanel"
                  aria-labelledby="settings-tab-advanced">
           <h3 id="settings-advanced-title" class="sr-only">Advanced</h3>
-          <label class="settings-field">Mobile touch compatibility
-            <select name="touchMode">
-              <option value="dbltap">Double-click sends double-tap</option>
-              <option value="translate">Translate mouse to touch</option>
-              <option value="augment">Mouse and touch together</option>
-              <option value="off">Off</option>
-            </select>
-          </label>
           <label class="settings-check">
             <input type="checkbox" name="showDiagnostics">
             <span>Show performance overlay</span>
